Refactored text utilities into `/src/utils/text.js`
