+++ src/components/decks/DeckDice.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import { capitalize } from '/src/utils.js'
+import { capitalize } from '/src/utils/text.js'
 
 export default {
   name: 'DeckDice',
